requirements.txt
google-generativeai>=0.8.0
mcp>=1.6.0
oracledb>=3.1.1
numpy>=1.24.0
rapidfuzz>=3.13.0
python-dotenv>=1.0.0